Develop

+++ src/main/java/com/model/GamesQuestions.java
@@ -1,7 +1,7 @@
 package trivia;
 import org.javalite.activejdbc.Model;
 public class GamesQuestions extends Model {
-	public GamesQuestions(Long quest_id, Long game_id){
+	public GamesQuestions(Integer quest_id, Integer game_id){
 		set("game_id",game_id);
 		set("question_id",quest_id);
 		saveIt();
